task-api/pyproject.toml
[project]
name = "task-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.27.0",
    "pytest-cov>=7.0.0",
]
